requirements.txt
transformers>=4.40.0
torch
faiss-cpu
pillow
numpy
tqdm
matplotlib
jupyter
